+++ NEWS.rst
@@ -4,6 +4,11 @@ News for stub
 This file lists the major changes between versions. For a more detailed list of
 every change, see the Git log.
 
+Latest
+------
+* Minor: Added the call_arguments(..) function to access a specific
+  call's arguments
+
 1.0.1
 -----
 * Bug: Add missing assert checking that the return_handler has actual